JWT + OAuth

+++ Controllers/OrdersController.cs
@@ -1,10 +1,12 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using RelicsAPI.Auth.Model;
 using RelicsAPI.Data.DTOs.Orders;
-using RelicsAPI.Data.DTOs.Relics;
 using RelicsAPI.Data.Entities;
 using RelicsAPI.Data.Repositories;
 
@@ -15,17 +17,19 @@ namespace RelicsAPI.Controllers
     public class OrdersController : ControllerBase
     {
         private readonly IOrdersRepository _ordersRepository;
-        private readonly IRelicsRepository _relicsRepository;
+        private readonly IAuthorizationService _authorizationService;
         private readonly IMapper _mapper;
 
-        public OrdersController(IOrdersRepository ordersRepository, IMapper mapper, IRelicsRepository relicsRepository)
+        public OrdersController(IOrdersRepository ordersRepository, IMapper mapper, IAuthorizationService authorizationService)
         {
             _ordersRepository = ordersRepository;
-            _relicsRepository = relicsRepository;
+            _authorizationService = authorizationService;
             _mapper = mapper;
         }
 
         [HttpGet]
+        [Route("all")]
+        [Authorize(Roles = UserRoles.Admin)]
         public async Task<IEnumerable<OrderDTO>> GetAll()
         {
             var orders = await _ordersRepository.GetAll();
@@ -33,7 +37,19 @@ namespace RelicsAPI.Controllers
             return orders.Select(o => _mapper.Map<OrderDTO>(o));
         }
 
+        [HttpGet]
+        [Authorize(Roles = UserRoles.Customer)]
+        public async Task<IEnumerable<OrderDTO>> GetAllCurrentUser()
+        {
+            var userId = (User.Identity as ClaimsIdentity).Claims.FirstOrDefault(c => c.Type == CustomClaims.UserId).Value;
+
+            var orders = await _ordersRepository.GetAllCurrentUser(userId);
+
+            return orders.Select(o => _mapper.Map<OrderDTO>(o));
+        }
+
         [HttpGet("{orderId}")]
+        [Authorize(Roles = UserRoles.Admin)]
         public async Task<ActionResult<OrderDTO>> GetById(int orderId)
         {
             var order = await _ordersRepository.GetById(orderId);
@@ -45,9 +61,13 @@ namespace RelicsAPI.Controllers
         }
         
         [HttpPost]
+        [Authorize(Roles = UserRoles.Customer)]
         public async Task<ActionResult<OrderDTO>> Create(CreateOrderDTO orderDTO)
         {
             var order = _mapper.Map<Order>(orderDTO);
+            var userId = (User.Identity as ClaimsIdentity).Claims.FirstOrDefault(c => c.Type == CustomClaims.UserId).Value;
+
+            order.UserId = userId;
 
             await _ordersRepository.Create(order);
 
@@ -55,12 +75,18 @@ namespace RelicsAPI.Controllers
         }
 
         [HttpPut("{orderId}")]
+        [Authorize(Roles = UserRoles.Customer)]
         public async Task<ActionResult<OrderDTO>> Update(int orderId, UpdateOrderDTO orderDTO)
         {
             var oldOrder = await _ordersRepository.GetById(orderId);
 
             if (oldOrder == null)
                 return NotFound($"Order with id {orderId} was not found");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, oldOrder, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+                return Forbid("Forbidden request"); // grazina 403, bet galima naudoti 404, kad neisduoti, jog toks resursas jau yra, 401 - kai blogas/nelegalus tokenas
 
             _mapper.Map(orderDTO, oldOrder);
 
@@ -70,12 +96,18 @@ namespace RelicsAPI.Controllers
         }
 
         [HttpDelete("{orderId}")]
+        [Authorize(Roles = UserRoles.Customer)]
         public async Task<ActionResult> Delete(int orderId)
         {
             var order = await _ordersRepository.GetById(orderId);
 
             if (order == null)
                 return NotFound($"Order with id {orderId} was not found");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, order, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+                return Forbid("Forbidden request"); // grazina 403, bet galima naudoti 404, kad neisduoti, jog toks resursas jau yra, 401 - kai blogas/nelegalus tokenas
 
             await _ordersRepository.Delete(order);
